% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd-package.R
\docType{package}
\name{grd-package}
\title{grd: Raster and Grid Geometry}
\description{
Provides a basic data structure for grids of points or rectangles including the math needed to index, extract, and resample these objects. A basic set of generics is provided such that downstream packages can extend the concept of a grid to alternative data sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/grd}
  \item Report bugs at \url{https://github.com/paleolimbot/grd/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
