% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generic.R
\name{grd_data_generic}
\alias{grd_data_generic}
\title{Wrap data sources with non-standard ordering}
\usage{
grd_data_generic(
  grid_data,
  data_order = grd_data_order(grid_data),
  ptype = grd_data_ptype(grid_data)
)
}
\arguments{
\item{grid_data}{The \code{data} member of a \code{\link[=grd]{grd()}}.}

\item{data_order}{A character vector with the
same length as \code{dim(grid_data)} specifying the
axis order and axis direction of indices in the
x y direction. The default \code{c("y", "x")} indicates
column-major ordering with y values decreasing
in the positive i index direction and x values increasing
in the positive j index direction. Use \code{"-y"} or \code{"-x"} to
switch axis directions. Use \code{NA} to indicate a non-xy
dimension.}

\item{ptype}{The R object type that should be used to represent
the data.}
}
\value{
An object of class \code{grd_data_generic}.
}
\description{
Wrap data sources with non-standard ordering
}
