% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{xy}
\alias{rct}
\alias{as_rct}
\alias{as_xy}
\alias{wk_bbox}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{wk}{\code{\link[wk:rct]{as_rct}}, \code{\link[wk:xy]{as_xy}}, \code{\link[wk]{rct}}, \code{\link[wk]{wk_bbox}}, \code{\link[wk]{xy}}}
}}

